SDI-525 Do not attempt to upsert prisoner hash if prisoner has not changed
This prevents a timing issue where multiple threads update the prisoner at the same time but no event is emitted. The sequence for the bug was:
* thread1 loads old prisoner details from ElasticSearch
* thread1 updates prisoner in ElasticSearch
* thread2 loads new prisoner details from ElasticSearch
* thread2 updates prisoner in ElasticSearch (with identical details)
* thread2 updates prisoner hash
* thread2 thinks old and new prisoner are identical so doesn't send update events
* thread1 tries to update prisoner hash but the hash was already updated by thread 2 so there is no change and thread1 also doesn't send update events
The fix is to prevent thread2 updating the prisoner hash - which is correct because it thinks the prisoner hasn't changed. This means when thread1 attempts to update the prisoner hash it succeeds and therefore sends the update events.